tests: Avoid side-effects in the settings files testing
We have many settings files which don't define options themselves, but
merely modify options inherited from their parent configuration.

For example, kb_bsfcampus.py adds some HOME_CARDS to the config it
inherits from its kb.py parent.

As a result, each call to test_setting_file would modify the same list
in memory. We never realized it because we never actually checked
such settings in this test, but it nevertheless was a problem.

This commit fixes it, ensuring that each call to the test method will
only get the settings from the tested file, and its parent(s).